--- a/scripts/generate-component.ts
+++ b/scripts/generate-component.ts
@@ -2,7 +2,6 @@
 
 import fs from 'fs';
 import path from 'path';
-import { execSync } from 'child_process';
 
 /**
  * Component Generator Script