… PROJECT_STATUS.md
… TODO.md
… ai-docs/test-driven-development-pytest.md
… ai-docs/test-driven-development-typescript.md