@raseniero 2025-08-08
✅ **Fixed**: Removed all superscript citation numbers from the TypeScript TDD documentation.

The file now flows naturally without the distracting , ,  etc. citations that were appearing as plain text. All technical content and methodology guidance remains intact.

**Changes made:**
- Removed 69+ superscript citation numbers throughout the document
- Maintained all technical content and TDD methodology 
- Improved readability by eliminating confusing text elements

**Commit**: 8a42a3b - Contains the complete fix addressing this review feedback.

Thanks for the catch! 🎯